fix: move file page
