+++ src/main.tsx
@@ -5,8 +5,8 @@ import './global.css'
 //CONFIG ROUTER
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-import Home from "../pages/Home/index";
-import Pokedex from "../pages/Pokedex/index"
+import Home from "./pages/Home/index";
+import Pokedex from "./pages/Pokedex/index"
 import App from './App';
 import ErrorPage from './components/ErrorPage';
